+++ contracts/sdk/gumball-machine/index.ts
@@ -2,3 +2,4 @@ export * from './instructions';
 export * from './accounts';
 export * from './types';
 export * from './utils';
+export * from './src/generated/index'
