+++ contracts/sdk/gumball-machine/package.json
@@ -1,7 +1,7 @@
 {
   "name": "gumball-machine",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "index.ts",
   "license": "MIT",
   "dependencies": {
     "@metaplex-foundation/rustbin": "^0.3.1",
